Remove omitempty from some fields.
This is a change to match the behaviour of the Fastly control panel.
When their CP interacts with their own API, they do send these fields in
when they are empty. Go-fastly does not. This behaviour difference means
that interacting with the control panel after using go-fastly, or
vice-versa, will result in apparent 'differences' between two versions
despite no changes being made.