Do not omitempty certain fields.
Fastly API has a bit of a quirk on certain object types. If you perform
an update to an object and don't include a field, it may zero out that
field, or it may consider it unchanged. This behaviour varies depending
on what object type you might be working with.

These fields, if not explicitly sent to Fastly with nil-like values,
will be considered unchanged by their API. As a result, if omitempty was
present here, it is impossible to zero out these fields.